requirements.txt
streamlit
Pillow
bpy